best_first_guess.c → best_guess.c
Allow distinct lists for allowed guesses and possible answers. Why?

1) Wordle does this (2,315 possible answers, but 12,972 allowed guesses,
   precisely matches CSW2019).  For how to extract and compare them, see
   https://github.com/dlenski/lexeme/#wordle-word-lists

2) After the first guess, solving for the best possible guess means
   using a larger set of allowed guesses than possible targets.

This also updates the solver to report the worst-case for each guess
("max_targets_left_after_guess") in addition to the average
("avg_targets_left_after_guess"), since it's been repeatedly and plausibly
conjectured that limiting the worst case is the better way to ensure a
successful solution within the allowed number of guesses.